+++ Repository/BiliUniverse/Global/CHANGELOG.md
@@ -1,3 +1,2 @@
-### 🔄 Other Changes
-  * 打包器由 `rollup` 更改为 `rspack`
-  * util 由 `submodule` 更改为 `package`
+### 🛠️ Bug Fixes
+  * 修复`Quantumult X`和`shadowrocket`兼容性